refactor: moved batching and types
