+++ op-service/client/batching_test.go
@@ -1,4 +1,4 @@
-package sources
+package client
 
 import (
 	"context"
