+++ op-service/client/types_test.go
@@ -1,4 +1,4 @@
-package sources
+package client
 
 import (
 	"encoding/json"
